requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
PyJWT==1.7.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
six==1.16.0
twilio==6.62.0
urllib3==1.26.6
